requirements.txt
numpy
tensorflow-cpu
pandas
matplotlib
sklearn
fastapi
uvicorn
pydantic
python-multipart
gunicorn
Jinja2